Lucid.Base

Description

Base types and combinators.

Rendering

renderText :: Html a -> Text

Render the HTML to a lazy Text.

This is a convenience function defined in terms of execHtmlT, runIdentity and toLazyByteString, and decodeUtf8. Check the source if you're interested in the lower-level behaviour.

renderBS :: Html a -> ByteString

Render the HTML to a lazy ByteString.

This is a convenience function defined in terms of execHtmlT, runIdentity and toLazyByteString. Check the source if you're interested in the lower-level behaviour.

renderTextT :: Monad m => HtmlT m a -> m Text

Render the HTML to a lazy Text, but in a monad.

This is a convenience function defined in terms of execHtmlT and toLazyByteString, and decodeUtf8. Check the source if you're interested in the lower-level behaviour.

renderBST :: Monad m => HtmlT m a -> m ByteString

Render the HTML to a lazy ByteString, but in a monad.

This is a convenience function defined in terms of execHtmlT and toLazyByteString. Check the source if you're interested in the lower-level behaviour.

renderToFile :: FilePath -> Html a -> IO ()

Render the HTML to a lazy ByteString.

This is a convenience function defined in terms of execHtmlT, runIdentity and toLazyByteString. Check the source if you're interested in the lower-level behaviour.

Running

execHtmlT

Arguments

:: Monad m

=> HtmlT m a
The HTML to generate.

-> m Builder
The a is discarded.

Build the HTML. Analogous to execState.

You might want to use this is if you want to do something with the raw Builder. Otherwise for simple cases you can just use renderText or renderBS.

evalHtmlT

Arguments

:: Monad m

=> HtmlT m a
HTML monad to evaluate.

-> m a
Ignore the HTML output and just return the value.

Evaluate the HTML to its return value. Analogous to evalState.

Use this if you want to ignore the HTML output of an action completely and just get the result.

For using with the Html type, you'll need runIdentity e.g.

>>> runIdentity (evalHtmlT (p_ "Hello!"))
()

runHtmlT :: HtmlT m a -> m (Seq Attribute -> Builder, a)

This is the low-level way to run the HTML transformer, finally returning an element builder and a value. You can pass mempty for this argument for a top-level call. See evalHtmlT and execHtmlT for easier to use functions.

relaxHtmlT

Arguments

:: Monad m

=> HtmlT Identity a
The HTML generated purely.

-> HtmlT m a
Same HTML accessible in a polymorphic context.

Generalize the underlying monad.

Some builders are happy to deliver results in a pure underlying monad, here Identity, but have trouble maintaining the polymorphic type. This utility generalizes from Identity.

Since: 2.9.6

commuteHtmlT

Arguments

:: (Functor m, Monad n)

=> HtmlT m a
unpurely generated HTML

-> m (HtmlT n a)
Commuted monads. Note: n can be Identity

Commute inner m to the front.

This is useful when you have impure HTML generation, e.g. using StateT. Recall, there is `MonadState s HtmlT` instance.

exampleHtml :: MonadState Int m => HtmlT m ()
exampleHtml = ul_ $ replicateM_ 5 $ do
  x <- get
  put (x + 1)
  li_ $ toHtml $ show x

exampleHtml' :: Monad m => HtmlT m ()
exampleHtml' = evalState (commuteHtmlT exampleHtml) 1

Since: 2.9.9

Combinators

makeElement

Arguments

:: Functor m

=> Text
Name.

-> HtmlT m a
Children HTML.

-> HtmlT m a
A parent element.

Make an HTML builder.

makeElementNoEnd

Arguments

:: Applicative m

=> Text
Name.

-> HtmlT m ()
A parent element.

Make an HTML builder for elements which have no ending tag.

makeXmlElementNoEnd

Arguments

:: Applicative m

=> Text
Name.

-> HtmlT m ()
A parent element.

Make an XML builder for elements which have no ending tag.

makeAttribute

Arguments

:: Text
Attribute name.

-> Text
Attribute value.

-> Attribute

Make an attribute builder.

Types

type Html = HtmlT Identity

Simple HTML builder type. Defined in terms of HtmlT. Check out that type for instance information.

Simple use-cases will just use this type. But if you want to transformer over Reader or something, you can go and use HtmlT.

newtype HtmlT m a

A monad transformer that generates HTML. Use the simpler Html type if you don't want to transform over some other monad.

Don't rely on the internal representation of this type. Use the monad and functor classes.

Constructors

HtmlT (m (Seq Attribute -> Builder, a))

data Attribute

A simple attribute. Don't use the constructor, use makeAttribute. Attributes are case sensitive, so if you want attributes to be merged properly, use a single case representation.

Constructors

Attribute !Text !Text

Classes

class Term arg result | result -> arg where

Used to construct HTML terms.

Simplest use: p_ = term "p" yields p_.

Very overloaded for three cases:

The first case is the basic arg of [(Text,Text)] which will return a function that wants children.

The second is an arg which is HtmlT m (), in which case the term accepts no attributes and just the children are used for the element.

Finally, this is also used for overloaded attributes, like style_ or title_. If a return type of (Text,Text) is inferred then an attribute will be made.

The instances look intimidating but actually the constraints make it very general so that type inference works well even in the presence of things like OverloadedLists and such.

Minimal complete definition

termWith

Methods

term

Arguments

:: Text
Name of the element or attribute.

-> arg
Either an attribute list or children.

-> result
Result: either an element or an attribute.

Used for constructing elements e.g. term "p" yields p_.

termWith

Arguments

:: Text
Name.

-> [Attribute]
Attribute transformer.

-> arg
Some argument.

-> result
Result: either an element or an attribute.

Use this if you want to make an element which inserts some pre-prepared attributes into the element.

class TermRaw arg result | result -> arg where

Same as the Term class, but will not HTML escape its children. Useful for elements like style_ or script_.

Minimal complete definition

termRawWith

Methods

termRaw

Arguments

:: Text
Name of the element or attribute.

-> arg
Either an attribute list or children.

-> result
Result: either an element or an attribute.

Used for constructing elements e.g. termRaw "p" yields p_.

termRawWith

Arguments

:: Text
Name.

-> [Attribute]
Attribute transformer.

-> arg
Some argument.

-> result
Result: either an element or an attribute.

Use this if you want to make an element which inserts some pre-prepared attributes into the element.

class ToHtml a where

Can be converted to HTML.

Methods

toHtml :: Monad m => a -> HtmlT m ()

Convert to HTML, doing HTML escaping.

toHtmlRaw :: Monad m => a -> HtmlT m ()

Convert to HTML without any escaping.

class With a where

With an element use these attributes. An overloaded way of adding attributes either to an element accepting attributes-and-children or one that just accepts attributes. See the two instances.

Methods

with

Arguments

:: a
Some element, either Html a or Html a -> Html a.

-> [Attribute]

-> a

With the given element(s), use the given attributes.